feat: connect auth flow to backend

+++ Projects/Features/AuthFeature/Sources/Data/DTO/SignupResponseDTO.swift
@@ -1,14 +1,14 @@
 public struct SignupResponseDTO: Codable, Equatable, Sendable {
-    public let id: Int
+    public let id: Int64
     public let email: String
     public let nickname: String
-    public let cashBalance: Int
+    public let cashBalance: Int64
 
     public init(
-        id: Int,
+        id: Int64,
         email: String,
         nickname: String,
-        cashBalance: Int
+        cashBalance: Int64
     ) {
         self.id = id
         self.email = email
