+++ Projects/Features/AuthFeature/Sources/Domain/Entity/AuthUser.swift
@@ -2,16 +2,16 @@
 ///
 /// 서버 응답 DTO가 아니라 Auth 도메인에서 의미 있는 사용자 모델이다.
 public struct AuthUser: Equatable, Sendable {
-    public let id: Int
+    public let id: Int64
     public let email: String
     public let nickname: String
-    public let cashBalance: Int
+    public let cashBalance: Int64
 
     public init(
-        id: Int,
+        id: Int64,
         email: String,
         nickname: String,
-        cashBalance: Int
+        cashBalance: Int64
     ) {
         self.id = id
         self.email = email
